TIL/04.01/ai-matching_v1/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
transformers>=4.31.0
timm>=0.9.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
requests>=2.28.0
python-dotenv>=1.0.0
tqdm>=4.65.0
clip @ git+https://github.com/openai/CLIP.git
salesforce-lavis>=1.0.2  # BLIP-2 model